restore pytest_plugin
