+++ jupyterlab_server/test_data/schemas/@jupyterlab/apputils-extension/themes.json
