+++ jupyterlab_server/test_data/schemas/@jupyterlab/codemirror-extension/commands.json
