+++ jupyterlab_server/test_data/schemas/@jupyterlab/shortcuts-extension/package.json.orig
