+++ jupyterlab_server/test_data/schemas/@jupyterlab/shortcuts-extension/plugin.json
