+++ jupyterlab_server/test_data/schemas/@jupyterlab/translation-extension/plugin.json
