+++ jupyterlab_server/test_data/schemas/@jupyterlab/unicode-extension/plugin.json
