+++ jupyterlab_server/test_data/workspaces/foo-2c26.jupyterlab-workspace
